+++ src/main/java/controller/HomeController.java
@@ -9,18 +9,13 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.net.URISyntaxException;
 
-public class HomeController extends AbstractController {
+public class HomeController extends Controller {
     private static final Logger logger = LoggerFactory.getLogger(HomeController.class);
 
     public static final String PATH = "/";
 
     @Override
-    void doPost(HttpRequest httpRequest, HttpResponse httpResponse) {
-
-    }
-
-    @Override
-    void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
+    protected void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
         try {
             httpResponse.setStatusCode(HttpStatus.OK);
             httpResponse.forward("/index.html");
